+++ CHANGELOG.rst
@@ -1,6 +1,6 @@
-2.0.0b2
-=======
-In Progress
+2.0.0
+=====
+May 28, 2014
 
 Features
 --------
@@ -30,6 +30,13 @@ Bug Fixes
 * Avoid submitting tasks to the ThreadPoolExecutor after shutdown. With
   retries enabled, this could cause Cluster.shutdown() to hang under
   some circumstances.
+* Fix unintended rebuild of token replica map when keyspaces are
+  discovered (on startup), added, or updated and TokenAwarePolicy is not
+  in use.
+* Avoid rebuilding token metadata when cluster topology has not
+  actually changed
+* Avoid preparing queries for hosts that should be ignored (such as
+  remote hosts when using the DCAwareRoundRobinPolicy) (PYTHON-75)
 
 Other
 ^^^^^